@@ -18,10 +18,10 @@
         "justinrainbow/json-schema": "~1.3.7",
         "php": ">=5.4",
         "twig/twig": "~1.16.0",
-        "symfony/config": "~2.5.0",
-        "symfony/dependency-injection": "~2.5.0",
-        "symfony/expression-language": "~2.5.0",
-        "symfony/filesystem": "~2.5.0",
-        "symfony/yaml": "~2.5.0"
+        "symfony/config": "~2.6.0",
+        "symfony/dependency-injection": "~2.6.0",
+        "symfony/expression-language": "~2.6.0",
+        "symfony/filesystem": "~2.6.0",
+        "symfony/yaml": "~2.6.0"
     }
 }
